+++ CHANGELOG.md
@@ -1,5 +1,17 @@
 # Changelog
 
+## 0.5.0 (2024-06-19)
+
+- Add `mobilerobotplanningclient`.
+
+## 0.4.0 (2024-06-19)
+
+- Divide the planning client and ZMQ client into separate repositories.
+
+## 0.3.7 (2024-06-02)
+
+- Allow finishStatus/finishMessage to be specified for task StopX commands.
+
 ## 0.3.6 (2024-04-30)
 
 - rename to "StartValidatePackFormation" to be more consistent with the threading model.